# Remove network initialization time from benchmark result