<!--- 👉👉👉 Provide a general summary of your changes. 👈👈👈 -->

## Motivation and Context
<!--- 💭💡💭 Why is this change required? What problem does it solve? 💭💡💭 -->
<!--- 🔗 If it fixes an open issue, please link to the issue here. 🔗 -->
We need the benchmark about one-shot of inference time, but our current benchmark script includes the time on network initialization. We should remove it.

## Description
<!--- 📝🎯📝 Describe your changes in detail. 📝🎯📝 -->
These are results of `LmnetV1Quantize`
#### FPGA result
* before 34msec
```
{
    "benchmark": {
        "inference": 0.03459699999999999, 
        "post": 0.01622699999999999, 
        "pre": 0.004686999999999997, 
        "total": 0.05551099999999998
    }, 
```
* after: 16msec
```
{
    "benchmark": {
        "inference": 0.016170999999999935, 
        "post": 0.01625399999999999, 
        "pre": 0.007635000000000058, 
        "total": 0.040059999999999985
    }, 
```

#### FPGA result
* before 52msec
```
{
    "benchmark": {
        "inference": 0.05205499999999996, 
        "post": 0.01645300000000005, 
        "pre": 0.004682999999999993, 
        "total": 0.073191
    }, 
```
* after: 36msec
```
{
    "benchmark": {
        "inference": 0.036363000000000034, 
        "post": 0.017635999999999985, 
        "pre": 0.007936999999999972, 
        "total": 0.06193599999999999
    }, 
```

## How has this been tested?
<!--- 🙆👌🙆 Please describe in detail how you tested your changes. 🙆👌🙆 -->
<!--- Include details of your testing environment, details of your manual tests, snippet code or commands of your manual tests, table of experiments results metrics, tests ran to see how your change affects other areas of the code, etc. -->
<!--- If you have experiments report documents please link to here. -->

## Screenshots (if appropriate):

## Types of changes
<!--- 🏁🎌🏁 What types of changes does your code introduce? Put an `x` in all the boxes that apply: 🏁🎌🏁 -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature / Optimization (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:
<!--- ✅✅✅ Go over all the following points, and put an `x` in all the boxes that apply. ✅✅✅ -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

<!--- TODO(wakisaka): After decide our code style, add this.
- [ ] My code follows the code style of this project.
-->
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.